+++ src/productos/dto/update-catalogo.dto.ts
