+++ src/productos/entities/catalogo.entity.ts
